feat: refactoring read_env to modular kubernetes services
