+++ app/v1/modules/kubernetes_services/__init__.py
@@ -2,8 +2,10 @@
 
 
 from .script_to_dict import ScriptToDictParserService
+from .script_read_env import ScriptEnvironReaderService
 
 
 __all__ = [
     "ScriptToDictParserService",
+    "ScriptEnvironReaderService",
 ]
